wis2bridge/requirements.txt
confluent-kafka
paho-mqtt
prometheus-client